feat(gui/react/feature): add priority raiser icon
An icon on the feature side list selected item, that allow
easy and visual increase of priorities. This approach has some
downsides, like when trying to increase the priority if an
item through a long list.

This could be improved with an enhanced "move to top" icon or
something like it.