refac(gui/react/feature): change selection property name
Concerns the selection of "details" from an item, and
not simply the selection of an item from the list. This
is important to differentiate between the intention to
navigate to the details of an item (when in narrow mode)
and simply selecting an item from the list, which should
have little to no impact whe in narrow mode.